requirements.txt
python-dotenv>=0.19.0
openai>=1.12.0 
tenacity>=8.0.1
click>=8.1.0